Adiciona CRUD Planejamento

+++ src/main/java/com/indtexbr/gestaonormas/controller/GestaoNormasController.java
@@ -1,13 +1,16 @@
 package com.indtexbr.gestaonormas.controller;
 
 import com.indtexbr.gestaonormas.model.Norma;
+import com.indtexbr.gestaonormas.model.Planejamento;
 import com.indtexbr.gestaonormas.repository.NormaRepository;
+import com.indtexbr.gestaonormas.repository.PlanejamentoRepository;
 import com.indtexbr.gestaonormas.service.FileStorageService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.websocket.server.PathParam;
 import java.io.IOException;
 import java.util.Optional;
 
@@ -21,20 +24,21 @@ public class GestaoNormasController {
     @Autowired
     private FileStorageService fileStorageService;
 
-    @GetMapping
-    @RequestMapping(path="/norma")
-    public Norma getNorma(@RequestParam (value = "id") Long id){
+    @Autowired
+    private PlanejamentoRepository planejamentoRepository;
+
+    @GetMapping("norma")
+    public Norma getNorma(@PathVariable ("id") Long id){
         Optional<Norma> opNorma = normaRepository.findById(id);
 
-        //if(opNorma.get() != null)
+        if(opNorma.get() != null)
             return opNorma.get(); //lança nullpointer exception?
 
-        //return null;
+        return null;
     }
 
-    @PostMapping
-    //@RequestMapping(path="/cadastranorma/")
-    public String criarNorma(@RequestParam ("file") MultipartFile file,
+    @PostMapping("norma")
+    public ResponseEntity criarNorma(@RequestParam ("file") MultipartFile file,
                              @RequestParam ("name") String name,
                              @RequestParam ("obs") String obs) throws IOException {
         Norma norma = new Norma();
@@ -48,16 +52,16 @@ public class GestaoNormasController {
             normaRepository.save(norma);
             message = "Uploaded the file successfully: " + file.getOriginalFilename();
 //            //return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
-            return message;
+            return new ResponseEntity(message, HttpStatus.OK);
         } catch (Exception e) {
             message = "Could not upload the file: " + file.getOriginalFilename() + "!";
 //            //return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponseMessage(message));
-            return message;
+            return new ResponseEntity(message, HttpStatus.BAD_REQUEST);
         }
     }
 
-    @PutMapping
-    public String atualizaNorma(@RequestParam ("id") Long id,
+    @PutMapping("norma")
+    public ResponseEntity atualizaNorma(@RequestParam ("id") Long id,
                                 @RequestParam (name = "file", required = false) MultipartFile file,
                                 @RequestParam (name = "name", required = false) String name,
                                 @RequestParam (name = "obs", required = false) String obs) throws IOException {
@@ -91,15 +95,62 @@ public class GestaoNormasController {
             }
 
             normaRepository.save(normaAtualizada);
-            message = "Norma Atualizada com sucesso";
+            return new ResponseEntity(HttpStatus.OK);
         }
 
-        return message;
+        return new ResponseEntity(id,HttpStatus.BAD_REQUEST);
     }
 
-    @DeleteMapping("{id}")
-    public String deletaNorma(@PathVariable("id")Long id){
-        normaRepository.deleteById(id);
-        return "ok";
+    @DeleteMapping("norma/{id}")
+    public ResponseEntity deletaNorma(@PathVariable("id")Long id){
+        Optional<Norma> normaCadastrada = normaRepository.findById(id);
+
+        if(!normaCadastrada.isEmpty())
+        {
+            normaRepository.deleteById(id);
+            return new ResponseEntity(normaCadastrada,HttpStatus.OK);
+        }
+
+        return new ResponseEntity(HttpStatus.BAD_REQUEST);
+    }
+
+    @GetMapping("planeja/{id}")
+    public ResponseEntity getPlanejamento(@PathVariable ("id") Long id){
+        Optional<Planejamento> optionalPlanejamento = planejamentoRepository.findById(id);
+
+        if(!optionalPlanejamento.isEmpty())
+        {
+            return new ResponseEntity(optionalPlanejamento.get(), HttpStatus.OK); //lança nullpointer exception?
+        }
+
+        return new ResponseEntity(id, HttpStatus.BAD_REQUEST);
+    }
+
+    @PostMapping("planeja")
+    public ResponseEntity criarPlanejamento(@RequestBody Planejamento planejamento) throws IOException {
+        return new ResponseEntity(planejamentoRepository.save(planejamento), HttpStatus.CREATED);
+    }
+
+    @PutMapping("planeja")
+    public ResponseEntity atualizaPlanejamento(@RequestBody Planejamento planejamento) throws IOException {
+        Optional<Planejamento> planejamentoCadastrado = planejamentoRepository.findById(planejamento.getId());
+
+        if(!planejamentoCadastrado.isEmpty())
+            return new ResponseEntity(planejamentoRepository.save(planejamento), HttpStatus.OK);
+
+        return new ResponseEntity(HttpStatus.BAD_REQUEST);
+    }
+
+    @DeleteMapping("planeja/{id}")
+    public ResponseEntity deletaPlanejamento(@PathVariable("id")Long id){
+        Optional<Planejamento> planejamentoCadastrado = planejamentoRepository.findById(id);
+
+        if(!planejamentoCadastrado.isEmpty())
+        {
+            planejamentoRepository.deleteById(id);
+            return new ResponseEntity(planejamentoCadastrado, HttpStatus.OK);
+        }
+
+        return new ResponseEntity(id,HttpStatus.BAD_REQUEST);
     }
 }
